pyproject.toml
[project]
name = "clip-vit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "faiss-cpu>=1.12.0",
    "huggingface-hub>=0.36.0",
    "mlx>=0.29.3",
    "mlx-lm>=0.28.3",
    "numpy>=2.3.4",
    "pillow>=11.3.0",
    "streamlit>=1.50.0",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "torchvision>=0.24.0",
    "transformers>=4.57.1",
]
